services/image/pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "image-service"
version = "0.1.0"
description = "Image Service for D&D Character Creator"
authors = ["Your Name <your.email@example.com>"]
packages = [{include = "image_service", from = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.103.0"
uvicorn = {extras = ["standard"], version = "^0.23.0"}
pydantic = {extras = ["email"], version = "^2.3.0"}
aio-pika = "^9.3.0"
httpx = "^0.24.1"
pillow = "^10.0.0"
minio = "^7.1.15"
python-multipart = "^0.0.6"
prometheus-client = "^0.17.0"
redis = {extras = ["hiredis"], version = "^5.0.0"}

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-asyncio = "^0.21.1"
pytest-cov = "^4.1.0"
black = "^23.7.0"
isort = "^5.12.0"
ruff = "^0.0.286"
mypy = "^1.5.1"

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3

[tool.mypy]
python_version = "3.11"
ignore_missing_imports = true
disallow_untyped_defs = true
check_untyped_defs = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_return_any = true
strict_optional = true
